1.3 Audio Manager
Created an audio manager for my game i can create arrays for the audio clips i want to have in my game.
next step is to make the button select the music and the needle to play the music.
created a Audiomanager script and a Sounds script
added 10 drum clips